device/requirements.txt
setuptools
paho-mqtt==1.4.0